+++ core/src/main/java/org/nvip/plugfest/tooling/sbom/uids/Hash.java
@@ -1,4 +1,4 @@
-package org.nvip.plugfest.tooling.sbom;
+package org.nvip.plugfest.tooling.sbom.uids;
 
 import java.util.Objects;